+++ scripts/src/actions/fix-yaml-config.ts
@@ -1,0 +1,12 @@
+import 'reflect-metadata'
+
+/*
+Put `github/*.yml` transform rules that you want executed during `Fix` workflow runs here.
+
+*Example*
+```ts
+import {protectDefaultBranches} from './shared/protect-default-branches'
+
+protectDefaultBranches()
+```
+*/